The backend now completely builds as a qml plugin.
