+++ input/sdleventloop.h
@@ -1,9 +1,13 @@
 #ifndef SDLEVENTLOOP_H
 #define SDLEVENTLOOP_H
 
-#include "joystick.h"
+class Joystick;
 
 // The SDLEventLoop's job is to poll for button states and to emit signals as devices are connected or disconnected.
+#include <QObject>
+#include <QTimer>
+#include <QMutex>
+#include <QHash>
 
 class SDLEventLoop : public QObject {
         Q_OBJECT
