+++ qmldir
@@ -1,4 +1,2 @@
-module PhxVideoOutput
-PhxVideoOutput 1.0
-plugin phxvideooutputplugin
-class VideoOutput
+module Phoenix.Backend
+plugin backend
