--- qmlplugins/coreplugin/qmldir
@@ -1,2 +1,0 @@
-module Phoenix.PhxCore
-plugin phxcoreplugin
