--- qmlplugins/inputplugin/phxinputplugin.h
@@ -1,14 +1,0 @@
-#ifndef PHXINPUTPLUGIN_H
-#define PHXINPUTPLUGIN_H
-
-#include <QQmlExtensionPlugin>
-
-class PhxInputPlugin : public QQmlExtensionPlugin
-{
-    Q_OBJECT
-    Q_PLUGIN_METADATA(IID "org.qt-project.Qt.QQmlExtensionInterface")
-public:
-    void registerTypes( const char *uri ) override;
-};
-
-#endif // INPUTPLUGIN_H
