--- qmlplugins/inputplugin/qmldir
@@ -1,2 +1,0 @@
-module Phoenix.PhxInput
-plugin phxinputplugin
