--- qmlplugins/qmlplugins.pro
@@ -1,8 +1,0 @@
-TEMPLATE = subdirs
-
-# We'll always be 64-bit
-CONFIG += x86_64
-
-SUBDIRS += inputplugin
-SUBDIRS += videoplugin
-SUBDIRS += coreplugin
