--- qmlplugins/videoplugin/qmldir
@@ -1,2 +1,0 @@
-module Phoenix.PhxVideoOutput
-plugin phxvideooutputplugin
