+++ util/logging.h
@@ -2,6 +2,7 @@
 #define LOGGING_H
 
 #include "backendcommon.h"
+#include <QDebug>
 
 /* This is used for debugging Phoenix. Instead of using qDebug(), developers should use qCDebug(%category%), such as
  * qCDebug(phxLibrary). The category used for debugging should be relevant to whatever class is being worked on.
@@ -16,6 +17,5 @@ Q_DECLARE_LOGGING_CATEGORY( phxControlProxy )
 Q_DECLARE_LOGGING_CATEGORY( phxVideo )
 Q_DECLARE_LOGGING_CATEGORY( phxAudioOutput )
 Q_DECLARE_LOGGING_CATEGORY( phxInput )
-Q_DECLARE_LOGGING_CATEGORY( phxLibrary )
 
 #endif
